requirements.txt
Flask
python-dotenv
Pandas
Numpy
scikit-learn